test: isolate agent access rendering coverage
